chore: some more Async changes, including default async ports. (#100)
* fix: remove "+ Sync" bound on "PollAsync" and "U" event bounds

As we only need events("U") to be "Send" for the channel and tokio only needs "Send" to spawn tasks.

* fix(listener): assign "taskpool" after creation

As otherwise "stop" will do nothing.
Somehow i missed this in #98.

* refactor(listener): have a "new" method and rely less on "start"

So that "start" can be skipped in the future when there are no sync tasks.

* feat(listener): add option to have a async ticker

Also now we can skip creating a sync-port worker if there are no sync tasks.

* fix(Cargo.toml): use feature "crossterm" instead of always including dependency "crossterm"

* fix(listener): change "PollAsync::poll" to take "&mut self"

* feat: add async crossterm input listener "CrosstermAsyncStream"

re #94

* test(builder): reduce amount of events checked

Due to flakeyness in CI (this time its the windows CI)

* chore(workflows/coverage): enable "async-ports" to also be tested