feat(utils): remove "get_title_or_center"
As commonly no title means nothing should be rendered.
Also the output basically only put into "get_block", which allows Options.
Finally, if a empty string is added as a title to a block, it is *always* drawn, even if that border side is not specified to be drawn.

BREAKING CHANGE:
"utils::get_title_or_center" has been removed